Rename main path
