+++ src/main/java/commands/PathHandler.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import exception.JessicaException;
-import frontend.Jessica;
+import jessica.Jessica;
 
 /**
  * Handles operations related to determining the storage path for the application.
