+++ src/main/java/jessica/Jessica.java
@@ -1,4 +1,4 @@
-package frontend;
+package jessica;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
