+++ src/main/java/jessica/Main.java
@@ -1,4 +1,4 @@
-package frontend;
+package jessica;
 
 import java.io.IOException;
