+++ src/main/resources/css/main.css
@@ -1,5 +1,5 @@
 .root {
-    main-color: rgb(237, 255, 242);       /* Create a looked-up color called "main-color" within root. */
+    main-color: #FEF9E1;       /* Create a looked-up color called "main-color" within root. */
     -fx-background-color: #FEF9E1;
 }
 
@@ -69,3 +69,12 @@
     -fx-background-color:orange;
     -fx-font-size: 20px;
 }
+
+.scroll-pane {
+    -fx-background-color: purple;
+}
+
+.scroll-pane .viewport {
+    -fx-background-color: transparent;
+}
+
